backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
mysqlclient==2.2.0
Pillow==10.1.0
djangorestframework-simplejwt==5.3.0

